feat(etapa-b): Products=catalog only, Inventory owns stock, scanner + transfers
## Schema
- Service.stock/minStock: Int → Decimal(12,4) — supports fractional quantities
- BranchInventory.stock/minStock: Int → Decimal(12,4)
- Migration: 20260529_inventory_stock_decimal.sql

## API — Inventory
- AdjustStockSchema: quantity Int → Float (supports fractions)
- TransferStockSchema: new (companyId, productId, fromBranchId, toBranchId, quantity, reason)
- InventoryService.transfer(): atomic tx — deducts from source, adds to dest, creates 2 TRANSFER movements, syncs global stock
- POST /inventory/transfer endpoint (INVENTORY_ADJUST permission)

## API — Products
- CreateProductSchema: stock/minStock use z.number() instead of z.number().int()

## Web — Products page
- Removed stock, minStock fields from create/edit form modal
- Removed Stock column from table → replaced with Inventario badge (Controlado/N/A)
- Removed lowStock filter button (belongs in Inventory module)
- Updated KPIs: removed valor inventario + stock bajo → shows total, with-stock count, categories, avg price
- Updated subtitle to clarify stock is in Inventory module
- trackStock switch kept (product attribute, not a quantity)

## Web — Inventory page
- Fixed products query: response is now {data,total} — extracts .data, limit:500
- Added useBarcodeScanner hook: fires when adjust modal is open, auto-fills productId by barcode/sku match
- Scanner toggle button in search bar with active indicator banner
- Barcode + SKU codes now visible in product table rows
- Stock values display decimals when fractional (2dp), integers otherwise
- Added branches query for transfer modal
- Transfer modal: product + from/to branch + quantity + reason
- Transfer button on each product row + global Transferir button in header
- Edit modal now includes minStock field (inventory module sets the alert threshold)
- Kardex columns: added Sucursal column to general kardex view

+++ apps/api/src/modules/tenant/products/products.service.ts
@@ -21,8 +21,8 @@ export const CreateProductSchema = z.object({
   uniMedida: z.number().int().default(59),
   // Stock
   trackStock: z.boolean().default(false),
-  stock:      z.number().int().nonnegative().optional().default(0),
-  minStock:   z.number().int().nonnegative().optional().default(0),
+  stock:      z.number().nonnegative().optional().default(0),
+  minStock:   z.number().nonnegative().optional().default(0),
   // Units / Fractions
   purchaseUnit:     z.string().max(50).optional().nullable(),
   saleUnitId:       z.string().cuid().optional().nullable(),
